TSQL job controller with feedback
The procedure checks whether another process has already called it (@Jobname);
if an instance of the procedure is already running, the current instance waits for the current one to complete.
A timeout can be set. You can also specify whether the process is restarted after the waiting time has elapsed.